+++ firmware/application/src/ble_main.h
@@ -6,6 +6,9 @@
 #include "ble_bas.h"
 
 
+extern uint16_t batt_lvl_in_milli_volts;
+extern uint8_t  percentage_batt_lvl;
+
 void ble_slave_init(void);
 void advertising_start(void);
 void nus_data_reponse(uint8_t *p_data, uint16_t length);