+++ firmware/application/src/data_cmd.h
@@ -30,6 +30,7 @@
 #define DATA_CMD_WIPE_FDS                       (1020)
 #define DATA_CMD_GET_ENABLED_SLOTS              (1023)
 #define DATA_CMD_DELETE_SLOT_SENSE_TYPE         (1024)
+#define DATA_CMD_GET_BATTERY_INFO               (1025)
 //
 // ******************************************************************